Add `get_shuffling` to state
