+++ eth2/types/Cargo.toml
@@ -9,9 +9,11 @@ bls = { path = "../utils/bls" }
 boolean-bitfield = { path = "../utils/boolean-bitfield" }
 ethereum-types = "0.4.0"
 hashing = { path = "../utils/hashing" }
+honey-badger-split =  { path = "../utils/honey-badger-split" }
 integer-sqrt = "0.1"
 rand = "0.5.5"
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0"
 ssz = { path = "../utils/ssz" }
+vec_shuffle = { path = "../utils/vec_shuffle" }
